# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Siyaho Printer Agent API
  version: 1.2.0
  description: Local bridge between web POS apps and network or Windows USB thermal printers. Open source — no license required.
servers:
  - url: http://127.0.0.1:17890
paths:
  /v1/health:
    get:
      summary: Agent health
      responses:
        '200':
          description: Agent is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  service:
                    type: string
                  version:
                    type: string
                  port:
                    type: integer
                  platform:
                    type: string
  /v1/printers:
    get:
      summary: List Windows printers (Windows agent only)
      responses:
        '200':
          description: Installed printers
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  printers:
                    type: array
                    items:
                      $ref: '#/components/schemas/WindowsPrinter'
        '501':
          description: Not supported on this platform
  /v1/print:
    post:
      summary: Print ESC/POS job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/LegacyDantsuPrint'
                - $ref: '#/components/schemas/V1Print'
      responses:
        '200':
          description: Print succeeded
        '400':
          description: Invalid request
        '500':
          description: Print failed
  /print:
    post:
      summary: Legacy print alias (DantSu)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegacyDantsuPrint'
      responses:
        '200':
          description: Print succeeded
components:
  schemas:
    WindowsPrinter:
      type: object
      properties:
        name:
          type: string
        portName:
          type: string
        driverName:
          type: string
        isDefault:
          type: boolean
    LegacyDantsuPrint:
      type: object
      required: [printer, payload]
      properties:
        printer:
          type: object
          properties:
            type:
              type: string
              enum: [network, windows]
            name:
              type: string
              description: Windows printer name (when type is windows)
            ip:
              type: string
            port:
              type: integer
              default: 9100
        payload:
          type: string
    V1Print:
      type: object
      properties:
        target:
          type: object
          properties:
            type:
              type: string
              enum: [network, windows]
            name:
              type: string
            host:
              type: string
            port:
              type: integer
        data:
          type: object
          properties:
            format:
              type: string
              enum: [dantsu, escpos]
            encoding:
              type: string
              example: base64
            content:
              type: string